Move category code to dttools/src/category.c
